ML developed with anom data from students that used FIES credit.
This is an academic work developed for Predictive Analysis.
I had to predict the defaul considering variables, which means we had to use supervised ML algorithms.

The best ML algorithms for solve this problem were Logistic Regression (crossvalidated and using LASSO L1) and Random Forrest.

Sensibility was the target score (95%) and one of the parameters were: wrong prediction costs much more when you say tha it wont happen default than the opposite.

I chose Random Forrest as best model for this prediction, removed the multi collinear variabels and had my score. Played a bit with the cutoff score to reach the target sensibility.